Add files via upload
General Python code for doing signal analysis. We don't have to use every file here but it might be good to just take a quick look at them and pick apart pieces that we need for our implementation. Using the FFT/oscope/GUI combination is going to probably be the biggest help.