Very very basic LPF and software amplifier
Low pass filter and software amp. 5 mins of work must be done to combine them and a days worth of work needs to be done to set the proper parameters for the filter. We need to create a sample audio file with human speech as well as some (preferably) high frequency rotor wash.